refactor(ts-client): move transaction execution into tx modules
Extract Seismic transaction execution into an explicit src/tx layer.

This moves the transparent send path, shielded send path, and signed
eth_call path into dedicated modules and centralizes shared Seismic tx
types under tx/types.ts. Wallet and contract internals now import from
the canonical tx modules directly, and the old root transaction files
have been removed.

Why this helps

The package previously hid most transaction behavior behind a small
number of overloaded root files. In practice, sendTransaction.ts was
acting as the core execution subsystem for multiple flows, which made it
hard to answer basic questions such as:
- where transparent writes are actually sent
- where shielded transaction preparation happens
- where signed eth_call execution lives
- which layers are routing decisions vs execution details

Making tx/ explicit gives transaction execution a clear home. That
improves readability for contributors, makes future refactors lower
risk, and gives the rest of the client a stable execution layer to build
on.

Architectural direction

The long-term direction is to separate the package by responsibility:
- client.ts defines client composition and the public surface
- actions/ defines client-facing methods and routing decisions
- contract/ contains ABI-aware helpers for reads and writes
- tx/ contains low-level execution paths for sends and calls
- extensions/ contains higher-level feature modules like deposit or SRC20

This change establishes tx/ as the canonical execution layer. That makes
later cleanup in actions/ and feature organization more straightforward
without changing user-facing behavior.

The refactor is structural only: it is intended to preserve existing
behavior while making it much clearer where send vs call execution lives.
JSDoc for the moved tx entrypoints is preserved in their new locations.
