+++ clients/ts/packages/seismic-viem/src/index.ts
@@ -31,7 +31,7 @@ export {
   transparentReadContract,
 } from '@sviem/contract/read.ts'
 export { estimateShieldedGas } from '@sviem/estimateGas.ts'
-export { signedCall } from '@sviem/signedCall.ts'
+export { signedCall } from '@sviem/tx/signedCall.ts'
 export type { ShieldedWriteContractDebugResult } from '@sviem/contract/write.ts'
 export {
   shieldedWriteContract,
